lab_01_02_02 replaced floats with doubles

+++ lab_01_02_02/main.c
@@ -4,18 +4,18 @@
 
 int main()
 {
-    float point1_x, point1_y;
-    float point2_x, point2_y;
-    float point3_x, point3_y;
+    double point1_x, point1_y;
+    double point2_x, point2_y;
+    double point3_x, point3_y;
 
-    scanf("%f%f", &point1_x, &point1_y);
-    scanf("%f%f", &point2_x, &point2_y);
-    scanf("%f%f", &point3_x, &point3_y);
+    scanf("%lf%lf", &point1_x, &point1_y);
+    scanf("%lf%lf", &point2_x, &point2_y);
+    scanf("%lf%lf", &point3_x, &point3_y);
 
-    float side1, side2, side3;
+    double side1, side2, side3;
 
-    float distance_x = point1_x - point2_x;
-    float distance_y = point1_y - point2_y;
+    double distance_x = point1_x - point2_x;
+    double distance_y = point1_y - point2_y;
     side1 = sqrt(distance_x * distance_x + distance_y * distance_y);
 
     distance_x = point2_x - point3_x;
@@ -26,9 +26,9 @@ int main()
     distance_y = point1_y - point3_y;
     side3 = sqrt(distance_x * distance_x + distance_y * distance_y);
 
-    float perimeter = side1 + side2 + side3;
+    double perimeter = side1 + side2 + side3;
 
-    printf("%f\n", perimeter);
+    printf("%lf\n", perimeter);
     
     return 0;
 }
